Modified To Implement Revenue Villages Census Details In Mandal Page and Alliance Parties Details In District Page  By Sivakumar

+++ partyanalyst-service/trunk/src/main/java/com/itgrids/partyanalyst/service/impl/StatePageService.java
@@ -205,7 +205,7 @@ public class StatePageService implements IStatePageService {
 				 censusVO.setMalePopulation(censusData.getTotalMalePopulation());
 				 censusVO.setFemalePopulation(censusData.getTotalFemalePopulation());
 				 censusVO.setTru(censusData.getTru());
-				 censusVO.setYear(censusData.getYear());
+				 censusVO.setYear(censusData.getYear().intValue());
 				 
 				 censusDataVO.add(censusVO);
 			 }
